+++ app/src/main/java/de/xikolo/controllers/base/NetworkStateFragment.kt
@@ -3,38 +3,18 @@ package de.xikolo.controllers.base
 
 import android.os.Bundle
 import android.view.*
-import androidx.annotation.StringRes
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 import de.xikolo.R
 import de.xikolo.controllers.helper.NetworkStateHelper
-import de.xikolo.extensions.observe
-import de.xikolo.network.jobs.base.NetworkCode
-import de.xikolo.utils.NetworkUtil
-import de.xikolo.utils.ToastUtil
-import de.xikolo.viewmodels.base.BaseViewModel
 
-abstract class NetworkStateFragment<T : BaseViewModel> : BaseFragment(), SwipeRefreshLayout.OnRefreshListener, ViewModelCreationInterface<T> {
+abstract class NetworkStateFragment : BaseFragment(), SwipeRefreshLayout.OnRefreshListener, NetworkStateHelper.NetworkStateOwner {
 
-    private lateinit var networkStateHelper: NetworkStateHelper
+    override lateinit var networkStateHelper: NetworkStateHelper
 
     abstract val layoutResource: Int
 
-    override lateinit var viewModel: T
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        initViewModel(this)
-    }
-
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        // inflate generic network state view
-        val networkStateView = inflater.inflate(R.layout.fragment_loading_state, container, false) as ViewGroup
-        // inflate content view inside
-        val contentView = networkStateView.findViewById<ViewStub>(R.id.content_view)
-        contentView.layoutResource = layoutResource
-        contentView.inflate()
-        // return complete view
-        return networkStateView
+        return inflateHelper(inflater, container, layoutResource)
     }
 
     override fun onCreateOptionsMenu(menu: Menu?, inflater: MenuInflater?) {
@@ -56,97 +36,6 @@ abstract class NetworkStateFragment<T : BaseViewModel> : BaseFragment(), SwipeRe
         super.onViewCreated(view, savedInstanceState)
 
         networkStateHelper = NetworkStateHelper(activity, view, this)
-
-        viewModel.networkState.observe(viewLifecycleOwner) {
-            if (it.code != NetworkCode.STARTED) hideAnyProgress()
-            when (it.code) {
-                NetworkCode.STARTED                   -> showAnyProgress()
-                NetworkCode.NO_NETWORK                -> if (it.userRequest || !networkStateHelper.contentViewVisible) showNetworkRequired()
-                NetworkCode.ERROR, NetworkCode.CANCEL -> showErrorMessage()
-                else                                  -> Unit
-            }
-        }
-
-        viewModel.onCreate()
-    }
-
-    override fun onRefresh() {
-        viewModel.onRefresh()
-    }
-
-    fun hideContent() {
-        networkStateHelper.hideContent()
-    }
-
-    fun showContent() {
-        networkStateHelper.showContent()
-    }
-
-    fun showBlockingProgress() {
-        networkStateHelper.showBlockingProgress()
-    }
-
-    fun showAnyProgress() {
-        networkStateHelper.showAnyProgress()
-    }
-
-    fun hideAnyProgress() {
-        networkStateHelper.hideAnyProgress()
-    }
-
-    fun showNetworkRequired() {
-        if (networkStateHelper.contentViewVisible) {
-            NetworkUtil.showNoConnectionToast()
-        } else {
-            networkStateHelper.setMessageTitle(R.string.notification_no_network)
-            networkStateHelper.setMessageSummary(R.string.notification_no_network_summary)
-            networkStateHelper.showMessage()
-        }
-    }
-
-    fun showLoginRequired(onClick: () -> Unit = {}) {
-        if (networkStateHelper.contentViewVisible) {
-            ToastUtil.show(R.string.toast_please_log_in)
-        } else {
-            networkStateHelper.setMessageTitle(R.string.notification_please_login)
-            networkStateHelper.setMessageSummary(R.string.notification_please_login_summary)
-            networkStateHelper.setMessageOnClickListener(View.OnClickListener { onClick() })
-            networkStateHelper.showMessage()
-        }
-    }
-
-    fun showErrorMessage() {
-        if (networkStateHelper.contentViewVisible) {
-            ToastUtil.show(R.string.error)
-        } else {
-            networkStateHelper.setMessageTitle(R.string.error)
-            networkStateHelper.setMessageSummary(null)
-            networkStateHelper.showMessage()
-        }
-    }
-
-    fun hideMessage() {
-        networkStateHelper.hideMessage()
-    }
-
-    fun showEmptyMessage(@StringRes title: Int, @StringRes summary: Int = 0) {
-        if (!networkStateHelper.anyProgressVisible) {
-            hideContent()
-            networkStateHelper.setMessageTitle(title)
-            networkStateHelper.setMessageSummary(summary)
-            networkStateHelper.showMessage()
-        }
-    }
-
-    fun showMessage(@StringRes title: Int, @StringRes summary: Int, onClick: () -> Unit = {}) {
-        if (networkStateHelper.contentViewVisible) {
-            ToastUtil.show(title)
-        } else {
-            networkStateHelper.setMessageTitle(title)
-            networkStateHelper.setMessageSummary(summary)
-            networkStateHelper.setMessageOnClickListener(View.OnClickListener { onClick() })
-            networkStateHelper.showMessage()
-        }
     }
 
 }
