+++ app/src/main/java/de/xikolo/controllers/course/CertificatesFragment.kt
@@ -8,7 +8,7 @@ import androidx.core.widget.TextViewCompat
 import butterknife.BindView
 import com.yatatsu.autobundle.AutoBundleField
 import de.xikolo.R
-import de.xikolo.controllers.base.NetworkStateFragment
+import de.xikolo.controllers.base.ViewModelFragment
 import de.xikolo.controllers.helper.DownloadViewHelper
 import de.xikolo.extensions.observe
 import de.xikolo.models.Course
@@ -17,7 +17,7 @@ import de.xikolo.models.Enrollment
 import de.xikolo.models.dao.EnrollmentDao
 import de.xikolo.viewmodels.course.CertificateListViewModel
 
-class CertificatesFragment : NetworkStateFragment<CertificateListViewModel>() {
+class CertificatesFragment : ViewModelFragment<CertificateListViewModel>() {
 
     companion object {
         val TAG: String = CertificatesFragment::class.java.simpleName
