+++ app/src/main/java/de/xikolo/controllers/course/ProgressFragment.kt
@@ -6,13 +6,13 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.yatatsu.autobundle.AutoBundleField
 import de.xikolo.R
-import de.xikolo.controllers.base.NetworkStateFragment
+import de.xikolo.controllers.base.ViewModelFragment
 import de.xikolo.extensions.observe
 import de.xikolo.models.dao.CourseProgressDao
 import de.xikolo.viewmodels.course.ProgressViewModel
 import de.xikolo.views.SpaceItemDecoration
 
-class ProgressFragment : NetworkStateFragment<ProgressViewModel>() {
+class ProgressFragment : ViewModelFragment<ProgressViewModel>() {
 
     companion object {
         val TAG: String = ProgressFragment::class.java.simpleName
