+++ app/src/main/java/de/xikolo/controllers/main/ViewModelMainFragment.kt
@@ -1,9 +1,10 @@
 package de.xikolo.controllers.main
 
-import de.xikolo.controllers.base.NetworkStateFragment
+import de.xikolo.R
+import de.xikolo.controllers.base.ViewModelFragment
 import de.xikolo.viewmodels.base.BaseViewModel
 
-abstract class ViewModelMainFragment<T : BaseViewModel> : NetworkStateFragment<T>() {
+abstract class ViewModelMainFragment<T : BaseViewModel> : ViewModelFragment<T>() {
 
     var activityCallback: MainActivityCallback? = null
 
@@ -19,6 +20,12 @@ abstract class ViewModelMainFragment<T : BaseViewModel> : NetworkStateFragment<T
     override fun onStop() {
         super.onStop()
         activityCallback = null
+    }
+
+    protected fun showLoginRequired() {
+        showLoginRequired {
+            activityCallback?.selectDrawerSection(R.id.navigation_login)
+        }
     }
 
 }
