--- app/src/main/java/de/xikolo/viewmodels/base/NullViewModel.kt
@@ -1,7 +1,0 @@
-package de.xikolo.viewmodels.base
-
-class NullViewModel : BaseViewModel() {
-
-    override fun onRefresh() {}
-
-}
